[논문] AI Agent 기반 Data Visualization 프레임워크 연구 (VisPath)/vispath/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apptools==5.3.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
bokeh==3.7.2
branca==0.8.1
brokenaxes==0.6.2
bs4==0.0.2
Cartopy==0.24.1
cdsapi==0.7.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cycler==0.12.1
dash==3.0.2
datapi==0.4.0
datasets==3.5.0
debugpy==1.8.13
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
envisage==7.0.3
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.0.3
folium==0.19.5
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
func_timeout==4.3.5
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
json5==0.12.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mayavi==4.8.2
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
multiurl==0.3.5
narwhals==1.33.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
openai==1.69.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
prettytable==3.16.0
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pyface==8.0.0
Pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
scipy==1.15.2
seaborn==0.13.2
selenium==4.31.0
setuptools==78.1.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.9.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
traitsui==8.0.0
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
vtk==9.4.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
win32_setctime==1.2.0
wsproto==1.2.0
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.19.0
zipp==3.21.0
